Cambio de ficheros sin acentos.
Intento de crear una búsqueda más eficiente de los palíndromos, comentada en el final del main.cpp

+++ VDinamico.h
@@ -23,7 +23,7 @@
 template <class T>
 class VDinamico{
     private:
-        long tamF, tamL;
+        int tamF, tamL;
         T* vector;
         
         void aumentarVector();
